# resources/openapi-mt.json
{
  "components": {
...
    "securitySchemes": {
      "api_key": {
        "bearerFormat": "AccessToken",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "dev@kamu.dev",
      "name": "Kamu Data Inc."
    },
    "description": "Decentralized data management tool",
    "license": {
      "name": ""
    },
    "title": "kamu-cli",
    "version": "0.205.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/{account_name}/{dataset_name}/blocks/{block_hash}": {
      "get": {
        "operationId": "dataset_blocks_handler",
        "parameters": [
          {
            "description": "Hash of the block",
            "in": "path",
            "name": "block_hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Get block by hash",
        "tags": [
          "odf-transfer"
        ]
      }
    },
    "/{account_name}/{dataset_name}/checkpoints/{physical_hash}": {
      "get": {
        "operationId": "dataset_checkpoints_get_handler",
        "parameters": [
          {
            "description": "Physical hash of the checkpoint",
            "in": "path",
            "name": "physical_hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Get checkpoint by hash",
        "tags": [
          "odf-transfer"
        ]
      },
      "put": {
        "operationId": "dataset_checkpoints_put_handler",
        "parameters": [
          {
            "description": "Physical hash of the checkpoint",
            "in": "path",
            "name": "physical_hash",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "items": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "default": null
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Upload checkpoint",
        "tags": [
          "odf-transfer"
        ]
      }
    },
    "/{account_name}/{dataset_name}/data/{physical_hash}": {
      "get": {
        "operationId": "dataset_data_get_handler",
        "parameters": [
          {
            "description": "Physical hash of the data slice",
            "in": "path",
            "name": "physical_hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Get data slice by hash",
        "tags": [
          "odf-transfer"
        ]
      },
      "put": {
        "operationId": "dataset_data_put_handler",
        "parameters": [
          {
            "description": "Physical hash of the data slice",
            "in": "path",
            "name": "physical_hash",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "items": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "default": null
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Upload data slice",
        "tags": [
          "odf-transfer"
        ]
      }
    },
...
    "/{account_name}/{dataset_name}/pull": {
      "get": {
        "operationId": "dataset_pull_ws_upgrade_handler",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "default": null
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Initiate pull via Smart Transfer Protocol",
        "tags": [
          "odf-transfer"
        ]
      }
    },
    "/{account_name}/{dataset_name}/push": {
      "get": {
        "operationId": "dataset_push_ws_upgrade_handler",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "default": null
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Initiate push via Smart Transfer Protocol",
        "tags": [
          "odf-transfer"
        ]
      }
    },
    "/{account_name}/{dataset_name}/refs/{reference}": {
      "get": {
        "operationId": "dataset_refs_handler",
        "parameters": [
          {
            "description": "Name of the reference",
            "in": "path",
            "name": "reference",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Get named block reference",
        "tags": [
          "odf-transfer"
        ]
      }
    },
...
  },
...
}